backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
ollama
pydantic
pymongo
motor
bcrypt
python-jose[cryptography]
reportlab
email-validator
